backend/pyproject.toml
[project]
name = "uni-lm"
version = "0.1.0"
description = "LLM orchestrator for students with intent classification and streaming responses"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.129.0",
    "ipython>=9.10.0",
    "langgraph>=1.0.8",
    "litellm>=1.81.12",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0",
    "redis>=7.2.0",
]

[project.optional-dependencies]
test = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.11.1",
    "httpx>=0.24.0",
]
dev = [
    "black>=23.7.0",
    "ruff>=0.0.285",
    "mypy>=1.4.1",
]

[tool.ruff]
src = ["backend/unilm/"]

[tool.ruff.lint]
ignore-init-module-imports = true


[tool.setuptools.packages.find]
where = ["."]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."] 